+++ engine/include/cowel/diagnostic.hpp
@@ -228,6 +228,9 @@ inline constexpr std::u8string_view to_str_zpad = u8"to_str.zpad";
 /// the given format is invalid.
 inline constexpr std::u8string_view to_str_format = u8"to_str.format";
 
+/// @brief In `\cowel_str_at`,
+/// the provided index is out of range.
+inline constexpr std::u8string_view str_at_range = u8"str.at.range";
 /// @brief In `\cowel_str_substr`,
 /// the provided start/length range is invalid.
 inline constexpr std::u8string_view str_substr_range = u8"str.substr.range";